revised main class to follow clean architecture principle
